feat: refactor graphql structure
