+++ src/graphql/mutations/root_mutation.ts
@@ -1,0 +1,3 @@
+const RootMutation = null;
+
+export default RootMutation;
