+++ src/graphql/types/case.ts
