+++ src/graphql/types/cpu.ts
