+++ src/graphql/types/graphic.ts
