+++ src/graphql/types/motherboard.ts
