+++ src/graphql/types/powersupplier.ts
