+++ src/graphql/types/ram.ts
